REF: Block.interpolate -> interpolate_2d_with_fill
